Add volumeType options for driver.
This patch introduce a new SC option called `volumeType`
for the user. This can be used to create volumes of desired type
based on the set parameter and value.

At present user is allowed to specify below volume types:

volumeType: replicate:6

or

volumeType: disperse:4:2

Where `4` is disperse data count and `2` is disperse redundancy count.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>