Properly preserve return code from gometalinter
Commit 4249bbd placed gometalinter into a pipeline when run in Travis to
output status as it proceeds. Unfortunately, this pipeline masked the
return code of gometalinter since it was not the last piped command.
This caused the linting to always pass. This fix sets the 'pipefail'
shell option to ensure non-zero exit status is preserved even when run
through the pipeline so that lint errors will properly return an error.

Signed-off-by: John Strunk <jstrunk@redhat.com>