Call Volume info instead of volume status API
After volume creation, suppose glusterd2
or CSI driver fails, as the volume is
not in the started state we won't be able to
get the volume status. as part of checking
the existing volume we don't need to call
volume status API, we can call volume info
and check volume is present or not.

Signed-off-by: Madhu Rajanna <mrajanna@redhat.com>