README: show the build status of the master branch
The current build status badge points to the smoke test that runs in the
CentOS CI. The result of the smoke tests for a PR does not tell anything
about the buildable status of the master branch.

A new job has been added to the CI to build the master branch after
changes have been merged. The updated link of the badge points to the
last build of the master branch now.

Fixes: #96
Signed-off-by: Niels de Vos <ndevos@redhat.com>